refactor(agent): one trace guard, not two (#606)
`requireRunTrace` (runner) and `requireExecutorTrace` (tool-executor)
read the same three fields through the same predicate and threw on the
same condition. Each file also carried its own byte-identical copy of
that predicate. Three helpers for one rule.

Now one `requireTrace(subject, traceContext)` beside `RunTrace`, whose
type it constructs, and one exported `nonEmptyString` beside it. The
subject names the boundary that refused, so the two messages stay
distinguishable.

One behavior change, an improvement: the tool executor's message was
"tool executor requires the run trace context" and is now "tool executor
requires a trace context with traceId, sessionId, runId" — the runner
already listed the missing fields, and a caller that forgot one should be
told which.
