+++ packages/agent/test/core/execution/tool-executor.test.ts
@@ -195,7 +195,7 @@ describe("createToolExecutor bus events", () => {
         toolExecutor: async () => ({ id: "r", toolCallId: "c", output: "" }),
         engine: PolicyEngine.create(),
       }),
-    ).toThrow("tool executor requires the run trace context");
+    ).toThrow("tool executor requires a trace context with");
   });
 
   it("publishes actor from trace context on PermissionDenied", async () => {
